+++ src/app/controllers/SessionController.js
@@ -1,22 +1,10 @@
 import jwt from 'jsonwebtoken';
-import * as Yup from 'yup';
 
 import authConfig from '../../config/auth';
 import User from '../models/User';
 
 class SessionController {
   async store(req, resp) {
-    const schema = Yup.object().shape({
-      email: Yup.string()
-        .email()
-        .required(),
-      password: Yup.string().required(),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return resp.status(400).json({ error: 'Validation fails' });
-    }
-
     const { email, password } = req.body;
 
     const user = await User.findOne({ where: { email } });
