+++ src/app/controllers/UserController.js
@@ -1,5 +1,3 @@
-import * as Yup from 'yup';
-
 import User from '../models/User';
 
 class UserController {
@@ -9,20 +7,6 @@ class UserController {
    *
    */
   async store(req, resp) {
-    const schema = Yup.object().shape({
-      name: Yup.string().required(),
-      email: Yup.string()
-        .email()
-        .required(),
-      password: Yup.string()
-        .min(6)
-        .required(),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return resp.status(400).json({ error: 'Validation fails' });
-    }
-
     const userExists = await User.findOne({ where: { email: req.body.email } });
     if (userExists) {
       return resp.status(400).json({ error: 'User already exists' });
@@ -39,24 +23,6 @@ class UserController {
    *
    */
   async update(req, resp) {
-    const schema = Yup.object().shape({
-      name: Yup.string(),
-      email: Yup.string().email(),
-      password: Yup.string().min(6),
-      oldPassword: Yup.string()
-        .min(6)
-        .when('password', (password, field) => {
-          return password ? field.required() : field;
-        }),
-      confirmPassword: Yup.string().when('password', (password, field) => {
-        return password ? field.required().oneOf([Yup.ref('password')]) : field;
-      }),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return resp.status(400).json({ error: 'Validation fails' });
-    }
-
     const { email, oldPassword } = req.body;
 
     const user = await User.findByPk(req.userId);
