+++ src/app/validators/MeetupStore.js
@@ -1,0 +1,21 @@
+import * as Yup from 'yup';
+
+export default async (req, resp, next) => {
+  try {
+    const schema = Yup.object().shape({
+      title: Yup.string().required(),
+      description: Yup.string().required(),
+      location: Yup.string().required(),
+      date: Yup.date().required(),
+      banner_id: Yup.number().required(),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+
+    return next();
+  } catch (err) {
+    return resp
+      .status(400)
+      .json({ error: 'Validation fails', messages: err.inner });
+  }
+};
