+++ src/app/validators/MeetupUpdate.js
@@ -1,0 +1,22 @@
+import * as Yup from 'yup';
+
+export default async (req, resp, next) => {
+  try {
+    const schema = Yup.object().shape({
+      title: Yup.string(),
+      description: Yup.string(),
+      location: Yup.string(),
+      date: Yup.date(),
+      banner_id: Yup.number(),
+      promoter_id: Yup.number(),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+
+    return next();
+  } catch (err) {
+    return resp
+      .status(400)
+      .json({ error: 'Validation fails', messages: err.inner });
+  }
+};
