+++ src/app/validators/UserStore.js
@@ -1,0 +1,23 @@
+import * as Yup from 'yup';
+
+export default async (req, resp, next) => {
+  try {
+    const schema = Yup.object().shape({
+      name: Yup.string().required(),
+      email: Yup.string()
+        .email()
+        .required(),
+      password: Yup.string()
+        .min(6)
+        .required(),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+
+    return next();
+  } catch (err) {
+    return resp
+      .status(400)
+      .json({ error: 'Validation fails', messages: err.inner });
+  }
+};
